fix: move `saveBaseValues` from skill_container `update` to skill `setContainer`
Having it in the `update` function of skill_container makes it inconsistent in terms of when the values are saved because it depends on the next `update` call. This is especially true when a skill is added during a state when the skill_container `IsUpdating` is true. This can lead to some unexpected behavior (I personally had to deal with such a case in PTR). Whereas having it in `setContainer` ensures that as soon as the skill is added via the skill_container `add` function, the base values are immediately saved.