feat: add ability to add skill events via MSU based `addEvent` function
The `_isEmpty` argument is there in case we want to implement a system later on where only those events are called which are present in a skill's nut file. In this case, non-empty functions should be always called and, therefore, it is necessary to know if the event is empty or not.