fix: call update after onUpdateLevel
This fixes issues in certain cases where if perks were added/removed or perk tree was modified during `onUpdateLevel` then the perk tree wasn't updated properly as the perks added were retained in the SkillsToAdd array of skill_container. An example of such a case is when gaining XP via Training Camp in the Legends mod and having a perk which adds perks in `onUpdateLevel`.