refactor!: remove getTotalArmorStaminaModifier from actor and add a substitute function in item_container
The new function accepts either an array of slots or a single slot. By default it returns the total StaminaModifier of items in all slots. It has built-in handling for items which may not have the getStaminaModifier() function in them.

The rationale behind removing the function from actor and adding to item_container is that StaminaModifier is a purely item property and while perks such as Brawny affect how much stamina the actor loses due to his items, the StaminaModifier of the items remains static. Therefore, keeping this function in the actor is misleading.