fix: refactor the query tooltip functions to include a call to the base vanilla function instead of complete overwrite
This fixes the issues caused by these functions having been updated in vanilla since the time the original hook in MSU was written leading to wrong behavior. This new method is less efficient as it calls getEntityById and getSkillByID while these things are also called by the vanilla functions, but this implementation is more robust against vanilla updates to these functions.