fix: add hooks on some ai behaviors to work properly with MSU item actions
This was necessary because in these ai behaviors the vanilla function calls `payForAction` without passing the swapped items to the `_items` array which triggers MSU's onPayForItemAction event without information about the items. This fix causes the vanilla's call to that `payForAction` to be ignored and then manually check for the changed items and call `payForAction` with those items.