feat: add ability to temporarily change weight until next roll
This allows situations where you want to have a single weighted container with default weights for a set of items. But you want to roll from it for different cases with different weight multipliers. Without such a system, this can only be accomplished by cloning the container, applying the multipliers, rolling from the clone and then discarding the clone every time you want to do this.