fix: unlock before setting the value and call with `_force`
This fixes the issue where the game was trying to set a locked setting's value during deserialization - which can potentially lead to the setting being locked to the wrong value (if it was locked both before loading the saved game and inside the saved game serialized state). It also led to the set function returning early and printing an error to the log, and never actually calling the callbacks for changing the value.