refactor: move includeFile, includeFiles and includeLoad functions to the MSU namespace
The old functions are converted to wrappers for the new MSU namespace functions. This is to help mods that may rely on these functions (e.g. Enduriel's mods) to function for now without requiring an update. These wrappers should be deleted in a future update after Enduriel updates his mods.