revert: the addition of _object parameter
This reverts a part of commit f151c037d3500b7f8f89c7df8a5e6724b086c2fe

This function should not change the values of any object - as this violates the single responsibility principle. It should take a metadata object and deserialize an object from it and return the deserialized object.