feat: set the AIBehaviorID of skills to vanilla behavior during create
Is accomplished via a hookDescendants on skill.nut which looks for the `create` function and then adds a custom function which is called at the end of the hooked `create`. The function uses a table which is populated during root_state.onInit function to match the skill id to its vanilla behavior id. This is done to avoid having to instantiate any file during queue, which breaks hooks.