fix: remove check for existing behavior
This is important because we add behavior by stack. And if we check for pre-existing behavior, then we don't add the behavior but we still remove it when the skill is removed. This leads to issues when a skill removes itself upon execution e.g. Throw Spear or Throw Net which get removed due to the item being one-use only. As this happens mid-execution of an ai behavior, the behavior is removed and the game freezing as the behavior is no longer attached to an actor.