fix: remove active behavior only after execution is complete
This fixes issues when a skill is used which removes itself upon execution, leading to the skill removing its behavior mid-execution of the behavior, which leads to errors and combat freeze. With this commit, we check if the removed skill's behavior is the currently active behavior, and if yes, then only remove it after the execution is complete.