fix: metadata emulators being empty and potentially causing issues
this would occur if a mod tried to use a property of metadata
they expected to exist.
For example if they use a Serialization.isSavedVersionAtLeast()
check on the emulated metadata this would error.
This method extract the correct information from world_state,
however it has obvious issues in that this is very non-standard behavior