fix: return via SkillNestedTooltipData for perks without perkdefs
This solves cases where a nested tooltip is requested for certain vanilla perks e.g. Stalwart with the key Perk+perk_stalwart. Because these perks exist as "perks" in the vanilla game but do not have defined perkdefs, so the PerkIDByFilename doesn't work for these. Instead, we do a workaround for them by treating them as regular skills.