feat: add NullEntitySkill tooltip key for nested tooltips
Sometimes you need to show the nested tooltip by considering the entity to be null. This is useful for e.g. showing nested tooltips of StatusEffects inside perk tooltips on the perk tree window when the selected character has the StatusEffect. Using the standard Skill+filename will show the tooltip of the effect based on that character whereas we want to show a generic tooltip independent of the entity.