fix: reset AdditionalAccuracy of skill before adding item's to it
Because an update already happens during addSkill, the AdditionalAccuracy of vanilla skills (e.g. Quick Shot) is set to that of the item due to the vanilla onAfterUpdate function of Quick Shot which sets its AdditionalAccuracy to that of the item. Then we added the item's accuracy to it again and then saved that as the base value. This led to doubling the AdditionalAccuracy in the skill. By resetting it before adding the item's we ensure that the item's AdditionalAccuracy is added to the base value of the skill that it had before the update loop.