fix: returning null if skill isn't present in NestedSkillItem
If NestedSkillItem was not null and the queried skill wasn't one of the skills provided by that item, then it would return null. With this fix it first checks if the nested skill item provides that skill, if not then it usese the SkillObjectsByFilename to use the default skill item.