refactor: set costs preview before calling skills function (#313)
This ensures that the actor's PreviewFatigue and PreviewActionPoints fields are properly set BEFORE the skill_container's onAffordabilityPreview function is called. However, this means that we need to update the UI again after calling onAffordabilityPreview (as the UI is initially updated in the call to the base setActiveEntityCostsPreview function). We could just call the `updateCostsPreview` on the JSHandle again manually, however that leads to somewhat glitchy animation in the UI as it tries to update twice. Therefore, we implement a custom table with delegate to disable the UI updating during the call to the base function, and then enable it after calling our onAffordabilityPreview event, after which we manually call the updateCostsPreview for the JSHandle update.