fix: FatigueCost getting wrong base value during equipping weapon
If the actor has a skill which modifies the weapon skills' FatigueCost field in e.g. onAfterUpdate function. This was happening because an update cycle happened during the call to the original `addSkill`, which also calls setFatigueCost and our hook on `setFatigueCost` was also setting the base value to the given cost. The latter is bad from a principle pov too because setting a skill's FatigueCost shouldn't modify its base fatigue cost.