fix: manually re-add Named itemtype instead of handling it differently
We re-add it directly in our hook on create because if someone overwrites setValuesBeforeRandomize we still want to ensure that the Named ItemType gets added back. We add in a check for isItemType for Named because some mod may want to remove it from certain items (in theory) so this is robust against that.