fix: only fire skill execution events if skill actually executed
Vanilla `use` function has checks built in which may cause it to return early and not actually execute the skill. Hooking onUse ensures that we only fire it when the skill is being executed, however in this case we still need a flag to switch from `use` as `onUse` can be called manually as well.