fix: allow null in writeString of serde emulator
When serializing a SerializationData, outside a regular save process of the game, the MetaData contains things such as Name, Filename etc. which are null but are stored in the SerializationEmulator with `writeString`. This does NOT throw any errors if the  SerializationData is serialized during a regular game save cycle using the native `_out` object instead of a SerializationEmulator. However, with the emulator it throws an exception due to this requirement. In order to match the behavior of the vanilla native object, we should allow nulls.