feat: re implement onMovementStep to be able to affect actor movement
This makes the function actually useful. It allows skills to stop an actor's movement at a particular tile based on conditions defined in the skill's `onMovementStep` function. Note that onMovementStep is called before moving to a tile. As long as all skills' onMovementStep returns true or null, the movement is valid. Otherwise:
- If even a single skill returns false, the movement is stopped at that tile prematurely.
- The skill can also choose to return a function which will then be executed and also cause the movement to stop. This function can be useful for example for refunding complete or partial movement costs or to trigger any other effects.

Note: Only the first skill that returns a movement stopping condition is considered. Therefore, skill order is important.