feat: add tables to facilitate translation of weapontypes and categories
The issue is that because the Key inside ::Const.Items.WeaponType is what is used to check for strings inside weapon.m.Categories to assign weapontype or vice versa, it creates a hurdle for translators as translated strings are not matched and therefore automatic weapontype assignment fails.

We change this by decoupling those strings and adding new tables where relevant strings can be pushed by translators.