fix: flip IsSaving and IsLoading during all BB scripts serde functions
People can use MSU SerDe Emulators to save/load BB tables outside the regular save/load process of the game. These onSerialize and onDeserialize functions of the BB tables may contain or may affect logic that is dependent on knowing whether the game is being loaded or saved. This means that said logic should not trigger while the object is being serialized/deserialized. Therefore, we need to flip the IsSaving and IsLoading to true during ser/de of all BB tables.